+++ src/utils/PushNotification.d.ts
@@ -1,7 +1,6 @@
-type PushNotification = {
+export const PushNotification: {
   register: (initApp: Function) => Promise<void>
   getToken: () => Promise<string>
   getVoipToken: () => Promise<string>
   resetBadgeNumber: () => void
-}
-export const PushNotification: PushNotification
+} = {}